AI News Summary And Mood Analysis Bot/requirements.txt
# Core dependencies
requests>=2.28
python-dotenv>=1.0.0
nltk>=3.8.1
vaderSentiment>=3.3.2
jinja2>=3.1.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas
matplotlib

# LLM providers (optional)
groq
openai>=1.0.0

# Google email integration (optional)
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
